requirements.txt
gradio==5.9.1
openai==1.59.3
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
tenacity==9.0.0
